#1667 Move prettyPrintStack to stack_out
